Rename application service classes
